refactor: PayValidator 단순 Pojo로 변겅 (@Service Annotation 제거)
